Move client-side Java code to `client` module.
Also moved some utility methods between classes.
