+++ client/src/main/java/org/spine3/Internal.java
