+++ client/src/main/java/org/spine3/base/package-info.java
