+++ client/src/main/java/org/spine3/io/file/FileNameEscaper.java
