+++ client/src/main/java/org/spine3/protobuf/error/MissingDescriptorException.java
