+++ client/src/main/java/org/spine3/protobuf/package-info.java
