+++ client/src/main/java/org/spine3/type/CommandClass.java
