+++ client/src/main/java/org/spine3/type/TypeName.java
