+++ client/src/main/java/org/spine3/type/package-info.java
