+++ client/src/main/java/org/spine3/util/Math.java
