+++ client/src/main/java/org/spine3/util/Tests.java
