+++ examples/src/main/java/org/spine3/examples/aggregate/Client.java
@@ -37,10 +37,10 @@ import java.util.List;
 
 import static com.google.common.collect.Lists.newArrayList;
 import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.spine3.client.ClientUtil.newUserId;
 import static org.spine3.examples.aggregate.ConnectionConstants.DEFAULT_CLIENT_SERVICE_PORT;
 import static org.spine3.examples.aggregate.Requests.*;
 import static org.spine3.protobuf.Messages.toText;
-import static org.spine3.util.Users.newUserId;
 
 /**
  * Sample of a client implementation.
