+++ examples/src/main/java/org/spine3/examples/aggregate/Requests.java
@@ -22,13 +22,13 @@ package org.spine3.examples.aggregate;
 import com.google.protobuf.Message;
 import org.spine3.base.CommandContext;
 import org.spine3.base.UserId;
+import org.spine3.client.ClientUtil;
 import org.spine3.client.CommandRequest;
 import org.spine3.examples.aggregate.command.AddOrderLine;
 import org.spine3.examples.aggregate.command.CreateOrder;
 import org.spine3.examples.aggregate.command.PayForOrder;
 import org.spine3.protobuf.Messages;
 import org.spine3.time.ZoneOffset;
-import org.spine3.util.Commands;
 
 /**
  * Utility class for generating sample command requests.
@@ -81,14 +81,14 @@ class Requests {
     //TODO:2016-01-14:alexander.yevsyukov: Use real utility method.
     public static CommandRequest newCommandRequest(UserId userId, Message command) {
         final CommandContext context = createCommandContext(userId);
-        final CommandRequest request = Commands.newCommandRequest(command, context);
+        final CommandRequest request = ClientUtil.newCommandRequest(command, context);
         return request;
     }
 
     //TODO:2016-01-14:alexander.yevsyukov: Use real method, which obtains TimeZone.
     public static CommandContext createCommandContext(UserId userId) {
         final CommandContext.Builder context = CommandContext.newBuilder()
-                .setCommandId(Commands.generateId())
+                .setCommandId(ClientUtil.generateId())
                 .setActor(userId)
                 .setZoneOffset(ZoneOffset.getDefaultInstance());
         return context.build();
