+++ server/src/test/java/org/spine3/testdata/TestEventRecordFactory.java
@@ -28,11 +28,11 @@ import org.spine3.test.project.event.ProjectCreated;
 import org.spine3.test.project.event.ProjectStarted;
 import org.spine3.test.project.event.TaskAdded;
 
+import static org.spine3.client.ClientUtil.newUserId;
 import static org.spine3.protobuf.Messages.toAny;
 import static org.spine3.testdata.TestAggregateIdFactory.createProjectId;
 import static org.spine3.testdata.TestContextFactory.createEventContext;
 import static org.spine3.testdata.TestEventFactory.*;
-import static org.spine3.util.Users.newUserId;
 
 /**
  * The utility class which is used for creating EventRecords for tests.
